requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
absl-py==2.3.0
    # via tensorboard
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
click==8.2.1
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via fastapi
fastapi==0.115.12
    # via aitt-symbol-clf (pyproject.toml)
fastapi-cli==0.0.7
    # via fastapi
filelock==3.18.0
    # via torch
fonttools==4.58.1
    # via matplotlib
fsspec==2025.5.1
    # via torch
grpcio==1.71.0
    # via tensorboard
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via fastapi
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
jinja2==3.1.6
    # via
    #   fastapi
    #   torch
joblib==1.5.1
    # via scikit-learn
kiwisolver==1.4.8
    # via matplotlib
lightning-utilities==0.14.3
    # via torchmetrics
markdown==3.8
    # via tensorboard
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
matplotlib==3.10.3
    # via aitt-symbol-clf (pyproject.toml)
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
networkx==3.5
    # via torch
numpy==2.2.6
    # via
    #   contourpy
    #   matplotlib
    #   opencv-python
    #   scikit-learn
    #   scipy
    #   tensorboard
    #   torchmetrics
    #   torchvision
opencv-python==4.11.0.86
    # via aitt-symbol-clf (pyproject.toml)
packaging==25.0
    # via
    #   lightning-utilities
    #   matplotlib
    #   tensorboard
    #   torchmetrics
pillow==11.2.1
    # via
    #   matplotlib
    #   torchvision
protobuf==6.31.1
    # via tensorboard
pydantic==2.11.5
    # via fastapi
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.1
    # via rich
pymupdf==1.26.0
    # via aitt-symbol-clf (pyproject.toml)
pyparsing==3.2.3
    # via matplotlib
python-dateutil==2.9.0.post0
    # via matplotlib
python-dotenv==1.1.0
    # via uvicorn
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.2
    # via uvicorn
rich==14.0.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.14.7
    # via fastapi-cli
scikit-learn==1.6.1
    # via aitt-symbol-clf (pyproject.toml)
scipy==1.15.3
    # via scikit-learn
setuptools==80.9.0
    # via
    #   lightning-utilities
    #   tensorboard
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   python-dateutil
    #   tensorboard
sniffio==1.3.1
    # via anyio
starlette==0.46.2
    # via fastapi
sympy==1.14.0
    # via torch
tensorboard==2.19.0
    # via aitt-symbol-clf (pyproject.toml)
tensorboard-data-server==0.7.2
    # via tensorboard
threadpoolctl==3.6.0
    # via scikit-learn
torch==2.7.0
    # via
    #   aitt-symbol-clf (pyproject.toml)
    #   torchmetrics
    #   torchvision
torchmetrics==1.7.2
    # via aitt-symbol-clf (pyproject.toml)
torchvision==0.22.0
    # via aitt-symbol-clf (pyproject.toml)
typer==0.16.0
    # via fastapi-cli
typing-extensions==4.13.2
    # via
    #   anyio
    #   fastapi
    #   lightning-utilities
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   torch
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
uvicorn==0.34.2
    # via
    #   fastapi
    #   fastapi-cli
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.5
    # via uvicorn
websockets==15.0.1
    # via uvicorn
werkzeug==3.1.3
    # via tensorboard
